pyproject.toml
[tool.poetry]
name = "meshtastic-bbs"
version = "1.0.0"
description = "This is a BBS system integrated with Meshtastic network. The system allows for message handling, bulletin boards, mail systems, and a channel directory."
authors = ["ricardodantas"]
license = "GPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pypubsub = "^4.0.3"
meshtastic = "^2.5.5"
pyfiglet = "^1.0.2"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.3"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
